requirements.txt
alembic==1.13.1
aniso8601==9.0.1
blinker==1.7.0
boto3==1.34.33
botocore==1.34.33
click==8.1.7
Flask==3.0.1
Flask-Cors==4.0.0
Flask-JWT-Extended==4.6.0
Flask-Mail==0.9.1
Flask-Migrate==4.0.5
Flask-RESTful==0.3.10
Flask-Reuploads==1.0.1.dev0
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
gunicorn==21.2.0
itsdangerous==2.1.2
Jinja2==3.1.3
jmespath==1.0.1
Mako==1.3.2
MarkupSafe==2.1.4
marshmallow==3.20.2
packaging==23.2
passlib==1.7.4
PyJWT==2.8.0
PyMySQL==1.1.0
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2023.4
s3transfer==0.10.0
six==1.16.0
SQLAlchemy==2.0.25
typing_extensions==4.9.0
urllib3==2.0.7
webargs==8.4.0
Werkzeug==3.0.1
